Move lookup logic into mixin
