+++ lib/brainstem/dsl/field.rb
@@ -1,6 +1,10 @@
+require 'brainstem/concerns/lookup'
+
 module Brainstem
   module DSL
     class Field
+      include Brainstem::Concerns::Lookup
+
       attr_reader :name, :type, :description, :conditionals, :options
 
       def initialize(name, type, description, options)
@@ -52,27 +56,6 @@ module Brainstem
         conditionals.all? { |conditional|
           presenter_conditionals[conditional].matches?(model, helper_instance, conditional_cache)
         }
-      end
-
-      private
-
-      def run_on_with_lookup(model, context, helper_instance)
-        context[:lookup][:fields][name] ||= begin
-          proc = options[:lookup]
-          lookup = helper_instance.instance_exec(context[:models], &proc)
-          if !options[:lookup_fetch].present? && !lookup.respond_to?(:[])
-            raise(StandardError, 'Brainstem expects the return result of the `lookup` to be a Hash since it must respond to [] in order to access the model\'s assocation(s). Default: lookup_fetch: lambda { |lookup, model| lookup[model.id] }`')
-          end
-
-          lookup
-        end
-
-        if options[:lookup_fetch]
-          proc = options[:lookup_fetch]
-          helper_instance.instance_exec(context[:lookup][:fields][name], model, &proc)
-        else
-          context[:lookup][:fields][name][model.id]
-        end
       end
     end
   end
